refactor: ErrorMessage 컴포넌트 단일화
- src/shared/components/error-message.tsx 제거 (form.tsx와 중복)
- src/shared/components/form.tsx의 ErrorMessage 유지하되 children 없을 때
  null 반환하도록 보완 (가드 없는 호출부 대응)
- import 4곳을 @/shared/components/form 으로 교체:
  - pages/auth/ui/{login-page, LoginPage, register-page}.tsx
  - features/create-group/components/CreateGroupForm.tsx

Refs: docs/refactoring/01-error-centralize.md
